+++ src/base/b-dynamic-page/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.193 (2021-05-28)
+
+#### :bug: Bug Fix
+
+* Fixed a bug when switching pages via `pageProp` and `page`
+
 ## v3.0.0-rc.192 (2021-05-27)
 
 #### :rocket: New Feature